+++ packages/transaction-controller/src/TransactionController.ts
@@ -13,7 +13,6 @@ import type {
 } from '@metamask/base-controller';
 import { BaseController } from '@metamask/base-controller';
 import {
-  safelyExecute,
   query,
   NetworkType,
   RPC,
@@ -46,12 +45,15 @@ import { updateGasFees } from './gasFees';
 import { addInitialHistorySnapshot, updateTransactionHistory } from './history';
 import { IncomingTransactionHelper } from './IncomingTransactionHelper';
 import { projectLogger as log } from './logger';
+import { pendingTransactionsLogger } from './logger';
+import { PendingTransactionTracker } from './PendingTransactionTracker';
 import { determineTransactionType } from './transaction-type';
 import type {
   DappSuggestedGasFees,
   TransactionParams,
   TransactionMeta,
   TransactionReceipt,
+  SecurityProviderRequest,
   SendFlowHistoryEntry,
   WalletDevice,
 } from './types';
@@ -71,7 +73,7 @@ import {
   validateIfTransactionUnapproved,
   validateMinimumIncrease,
 } from './utils';
-import { validateTxParams } from './validation';
+import { validateTransactionOrigin, validateTxParams } from './validation';
 
 export const HARDFORK = Hardfork.London;
 
@@ -98,12 +100,10 @@ export interface FeeMarketEIP1559Values {
  * @type TransactionConfig
  *
  * Transaction controller configuration
- * @property interval - Polling interval used to fetch new currency rate
  * @property provider - Provider used to create a new underlying EthQuery instance
  * @property sign - Method used to sign transactions
  */
 export interface TransactionConfig extends BaseConfig {
-  interval: number;
   sign?: (txParams: TransactionParams, from: string) => Promise<any>;
   txHistoryLimit: number;
 }
@@ -183,7 +183,7 @@ export class TransactionController extends BaseController<
 
   private readonly provider: Provider;
 
-  private handle?: ReturnType<typeof setTimeout>;
+  private readonly handle?: ReturnType<typeof setTimeout>;
 
   private readonly mutex = new Mutex();
 
@@ -195,9 +195,17 @@ export class TransactionController extends BaseController<
 
   private readonly getGasFeeEstimates: () => Promise<GasFeeState>;
 
+  private readonly getPermittedAccounts: (origin?: string) => Promise<string[]>;
+
+  private readonly getSelectedAddress: () => string;
+
   private readonly messagingSystem: TransactionControllerMessenger;
 
   private readonly incomingTransactionHelper: IncomingTransactionHelper;
+
+  private readonly securityProviderRequest?: SecurityProviderRequest;
+
+  private readonly pendingTransactionTracker: PendingTransactionTracker;
 
   private failTransaction(transactionMeta: TransactionMeta, error: Error) {
     const errorData = {
@@ -258,6 +266,7 @@ export class TransactionController extends BaseController<
    * @param options.getCurrentNetworkEIP1559Compatibility - Whether or not the network supports EIP-1559.
    * @param options.getGasFeeEstimates -
    * @param options.getNetworkState - Gets the state of the network controller.
+   * @param options.getPermittedAccounts - Get accounts that a given origin has permissions for.
    * @param options.getSelectedAddress - Gets the address of the currently selected account.
    * @param options.incomingTransactions - Configuration options for incoming transaction support.
    * @param options.incomingTransactions.includeTokenTransfers - Whether or not to include ERC20 token transfers.
@@ -267,6 +276,7 @@ export class TransactionController extends BaseController<
    * @param options.messenger - The controller messenger.
    * @param options.onNetworkStateChange - Allows subscribing to network controller state changes.
    * @param options.provider - The provider used to create the underlying EthQuery instance.
+   * @param options.securityProviderRequest - A function for verifying a transaction, whether it is malicious or not.
    * @param config - Initial options used to configure this controller.
    * @param state - Initial state to set on this controller.
    */
@@ -279,11 +289,13 @@ export class TransactionController extends BaseController<
       getCurrentNetworkEIP1559Compatibility,
       getGasFeeEstimates,
       getNetworkState,
+      getPermittedAccounts,
       getSelectedAddress,
       incomingTransactions = {},
       messenger,
       onNetworkStateChange,
       provider,
+      securityProviderRequest,
     }: {
       blockTracker: BlockTracker;
       disableHistory: boolean;
@@ -292,6 +304,7 @@ export class TransactionController extends BaseController<
       getCurrentNetworkEIP1559Compatibility: () => Promise<boolean>;
       getGasFeeEstimates: () => Promise<GasFeeState>;
       getNetworkState: () => NetworkState;
+      getPermittedAccounts: (origin?: string) => Promise<string[]>;
       getSelectedAddress: () => string;
       incomingTransactions: {
         includeTokenTransfers?: boolean;
@@ -302,6 +315,7 @@ export class TransactionController extends BaseController<
       messenger: TransactionControllerMessenger;
       onNetworkStateChange: (listener: (state: NetworkState) => void) => void;
       provider: Provider;
+      securityProviderRequest?: SecurityProviderRequest;
     },
     config?: Partial<TransactionConfig>,
     state?: Partial<TransactionState>,
@@ -309,7 +323,6 @@ export class TransactionController extends BaseController<
     super(config, state);
 
     this.defaultConfig = {
-      interval: 15000,
       txHistoryLimit: 40,
     };
 
@@ -324,6 +337,7 @@ export class TransactionController extends BaseController<
     this.provider = provider;
     this.messagingSystem = messenger;
     this.getNetworkState = getNetworkState;
+    // @ts-expect-error TODO: Provider type alignment
     this.ethQuery = new EthQuery(provider);
     this.isSendFlowHistoryDisabled = disableSendFlowHistory ?? false;
     this.isHistoryDisabled = disableHistory ?? false;
@@ -333,6 +347,9 @@ export class TransactionController extends BaseController<
     this.getCurrentNetworkEIP1559Compatibility =
       getCurrentNetworkEIP1559Compatibility;
     this.getGasFeeEstimates = getGasFeeEstimates;
+    this.getPermittedAccounts = getPermittedAccounts;
+    this.getSelectedAddress = getSelectedAddress;
+    this.securityProviderRequest = securityProviderRequest;
 
     this.nonceTracker = new NonceTracker({
       provider,
@@ -375,26 +392,32 @@ export class TransactionController extends BaseController<
       this.onUpdatedLastFetchedBlockNumbers.bind(this),
     );
 
+    this.pendingTransactionTracker = new PendingTransactionTracker({
+      blockTracker,
+      failTransaction: this.failTransaction.bind(this),
+      getChainId: this.getChainId.bind(this),
+      getEthQuery: () => this.ethQuery,
+      getTransactions: () => this.state.transactions,
+    });
+
+    this.pendingTransactionTracker.hub.on(
+      'transactions',
+      this.onPendingTransactionsUpdate.bind(this),
+    );
+
+    this.pendingTransactionTracker.hub.on(
+      'transaction-confirmed',
+      (transactionMeta: TransactionMeta) =>
+        this.hub.emit(`${transactionMeta.id}:confirmed`, transactionMeta),
+    );
+
     onNetworkStateChange(() => {
+      // @ts-expect-error TODO: Provider type alignment
       this.ethQuery = new EthQuery(this.provider);
       this.registry = new MethodRegistry({ provider: this.provider });
     });
 
-    this.poll();
-  }
-
-  /**
-   * Starts a new polling interval.
-   *
-   * @param interval - The polling interval used to fetch new transaction statuses.
-   */
-  async poll(interval?: number): Promise<void> {
-    interval && this.configure({ interval }, false, false);
-    this.handle && clearTimeout(this.handle);
-    await safelyExecute(() => this.queryTransactionStatuses());
-    this.handle = setTimeout(() => {
-      this.poll(this.config.interval);
-    }, this.config.interval);
+    this.pendingTransactionTracker.start();
   }
 
   /**
@@ -432,6 +455,7 @@ export class TransactionController extends BaseController<
    * @param opts - Additional options to control how the transaction is added.
    * @param opts.actionId - Unique ID to prevent duplicate requests.
    * @param opts.deviceConfirmedOn - An enum to indicate what device confirmed the transaction.
+   * @param opts.method - RPC method that requested the transaction.
    * @param opts.origin - The origin of the transaction request, such as a dApp hostname.
    * @param opts.requireApproval - Whether the transaction requires approval by the user, defaults to true unless explicitly disabled.
    * @param opts.securityAlertResponse - Response from security validator.
@@ -447,6 +471,7 @@ export class TransactionController extends BaseController<
     {
       actionId,
       deviceConfirmedOn,
+      method,
       origin,
       requireApproval,
       securityAlertResponse,
@@ -456,6 +481,7 @@ export class TransactionController extends BaseController<
     }: {
       actionId?: string;
       deviceConfirmedOn?: WalletDevice;
+      method?: string;
       origin?: string;
       requireApproval?: boolean | undefined;
       securityAlertResponse?: Record<string, unknown>;
@@ -469,6 +495,14 @@ export class TransactionController extends BaseController<
     txParams = normalizeTxParams(txParams);
     const isEIP1559Compatible = await this.getEIP1559Compatibility();
     validateTxParams(txParams, isEIP1559Compatible);
+    if (origin) {
+      await validateTransactionOrigin(
+        await this.getPermittedAccounts(origin),
+        this.getSelectedAddress(),
+        txParams.from,
+        origin,
+      );
+    }
 
     const dappSuggestedGasFees = this.generateDappSuggestedGasFees(
       txParams,
@@ -514,6 +548,15 @@ export class TransactionController extends BaseController<
 
     // Checks if a transaction already exists with a given actionId
     if (!existingTransactionMeta) {
+      // Set security provider response
+      if (method && this.securityProviderRequest) {
+        const securityProviderResponse = await this.securityProviderRequest(
+          transactionMeta,
+          method,
+        );
+        transactionMeta.securityProviderResponse = securityProviderResponse;
+      }
+
       if (!this.isSendFlowHistoryDisabled) {
         transactionMeta.sendFlowHistory = sendFlowHistory ?? [];
       }
@@ -851,37 +894,6 @@ export class TransactionController extends BaseController<
       estimateGasError,
       simulationFails,
     };
-  }
-
-  /**
-   * Check the status of submitted transactions on the network to determine whether they have
-   * been included in a block. Any that have been included in a block are marked as confirmed.
-   */
-  async queryTransactionStatuses() {
-    const { transactions } = this.state;
-    const currentChainId = this.getChainId();
-    let gotUpdates = false;
-    await safelyExecute(() =>
-      Promise.all(
-        transactions.map(async (meta, index) => {
-          if (!meta.verifiedOnBlockchain && meta.chainId === currentChainId) {
-            const [reconciledTx, updateRequired] =
-              await this.blockchainTransactionStateReconciler(meta);
-            if (updateRequired) {
-              transactions[index] = reconciledTx;
-              gotUpdates = updateRequired;
-            }
-          }
-        }),
-      ),
-    );
-
-    /* istanbul ignore else */
-    if (gotUpdates) {
-      this.update({
-        transactions: this.trimTransactionsForState(transactions),
-      });
-    }
   }
 
   /**
@@ -1414,101 +1426,6 @@ export class TransactionController extends BaseController<
     ].includes(status);
   }
 
-  /**
-   * Method to verify the state of a transaction using the Blockchain as a source of truth.
-   *
-   * @param meta - The local transaction to verify on the blockchain.
-   * @returns A tuple containing the updated transaction, and whether or not an update was required.
-   */
-  private async blockchainTransactionStateReconciler(
-    meta: TransactionMeta,
-  ): Promise<[TransactionMeta, boolean]> {
-    const { status, hash } = meta;
-    switch (status) {
-      case TransactionStatus.confirmed:
-        const txReceipt = await query(this.ethQuery, 'getTransactionReceipt', [
-          hash,
-        ]);
-
-        if (!txReceipt) {
-          return [meta, false];
-        }
-
-        const txBlock = await query(this.ethQuery, 'getBlockByHash', [
-          txReceipt.blockHash,
-        ]);
-
-        meta.verifiedOnBlockchain = true;
-        meta.txParams.gasUsed = txReceipt.gasUsed;
-        meta.txReceipt = txReceipt;
-        meta.baseFeePerGas = txBlock?.baseFeePerGas;
-        meta.blockTimestamp = txBlock?.timestamp;
-
-        // According to the Web3 docs:
-        // TRUE if the transaction was successful, FALSE if the EVM reverted the transaction.
-        if (Number(txReceipt.status) === 0) {
-          const error: Error = new Error(
-            'Transaction failed. The transaction was reversed',
-          );
-          this.failTransaction(meta, error);
-          return [meta, false];
-        }
-
-        return [meta, true];
-      case TransactionStatus.submitted:
-        const txObj = await query(this.ethQuery, 'getTransactionByHash', [
-          hash,
-        ]);
-
-        if (!txObj) {
-          const receiptShowsFailedStatus =
-            await this.checkTxReceiptStatusIsFailed(hash);
-
-          // Case the txObj is evaluated as false, a second check will
-          // determine if the tx failed or it is pending or confirmed
-          if (receiptShowsFailedStatus) {
-            const error: Error = new Error(
-              'Transaction failed. The transaction was dropped or replaced by a new one',
-            );
-            this.failTransaction(meta, error);
-          }
-        }
-
-        /* istanbul ignore next */
-        if (txObj?.blockNumber) {
-          meta.status = TransactionStatus.confirmed;
-          this.hub.emit(`${meta.id}:confirmed`, meta);
-          return [meta, true];
-        }
-
-        return [meta, false];
-      default:
-        return [meta, false];
-    }
-  }
-
-  /**
-   * Method to check if a tx has failed according to their receipt
-   * According to the Web3 docs:
-   * TRUE if the transaction was successful, FALSE if the EVM reverted the transaction.
-   * The receipt is not available for pending transactions and returns null.
-   *
-   * @param txHash - The transaction hash.
-   * @returns Whether the transaction has failed.
-   */
-  private async checkTxReceiptStatusIsFailed(
-    txHash: string | undefined,
-  ): Promise<boolean> {
-    const txReceipt = await query(this.ethQuery, 'getTransactionReceipt', [
-      txHash,
-    ]);
-    if (!txReceipt) {
-      // Transaction is pending
-      return false;
-    }
-    return Number(txReceipt.status) === 0;
-  }
-
   private async requestApproval(
     txMeta: TransactionMeta,
     { shouldShowRequest }: { shouldShowRequest: boolean },
@@ -1636,6 +1553,11 @@ export class TransactionController extends BaseController<
   }) {
     this.update({ lastFetchedBlockNumbers });
     this.hub.emit('incomingTransactionBlock', blockNumber);
+  }
+
+  private onPendingTransactionsUpdate(transactions: TransactionMeta[]) {
+    pendingTransactionsLogger('Updated pending transactions');
+    this.update({ transactions: this.trimTransactionsForState(transactions) });
   }
 
   private generateDappSuggestedGasFees(