+++ packages/transaction-controller/src/logger.ts
@@ -1,3 +1,5 @@
+/* istanbul ignore file */
+
 import { createProjectLogger, createModuleLogger } from '@metamask/utils';
 
 export const projectLogger = createProjectLogger('transaction-controller');
@@ -5,6 +7,11 @@ export const projectLogger = createProjectLogger('transaction-controller');
 export const incomingTransactionsLogger = createModuleLogger(
   projectLogger,
   'incoming-transactions',
+);
+
+export const pendingTransactionsLogger = createModuleLogger(
+  projectLogger,
+  'pending-transactions',
 );
 
 export { createModuleLogger };