+++ packages/polling-controller/CHANGELOG.md
@@ -6,4 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/MetaMask/core/
+## [0.1.0]
+### Added
+- Initial release
+
+[Unreleased]: https://github.com/MetaMask/core/compare/@metamask/polling-controller@0.1.0...HEAD
+[0.1.0]: https://github.com/MetaMask/core/releases/tag/@metamask/polling-controller@0.1.0